+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Srdimport: Check for discrepancy between the HP in the Excel file and the calculated value
+- Srdimport: output log in CSV format to import into spreadsheet for analysis
+- Srdimport: add list of imported monsters at the end of the log file, with a section for warnings
 
 ### Changed
 - Srdimport: improve calculation of shape